backend/requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30
python-whois>=0.9.5
joblib>=1.4
scikit-learn>=1.5
numpy>=1.26
opencv-python-headless>=4.10
pandas>=2.2
python-multipart>=0.0.9
